+++ tests/testthat/test_fhir_collapse.R
@@ -1,6 +1,3 @@
-brackets <- c("[", "]")
-sep <- "|"
-
 testthat::test_that(
 	"fhir_collapse works on address.line",{
 		expect_snapshot_value({
@@ -12,8 +9,7 @@ testthat::test_that(
 							sep = sep),
 						verbose = 0
 		)
-
-		fhir_collapse(d, columns = "address.line", sep = sep, brackets = brackets)
+		fhir_collapse(d, columns = "address.line", sep = "|", brackets = c("[", "]"))
 		}, style = "json2")
 
 	})
@@ -25,11 +21,11 @@ testthat::test_that(
 		d <- fhir_crack(bundles,
 						design = fhir_table_description(
 							resource = "Patient",
-							brackets = brackets,
-							sep = sep),
+							brackets = c("[", "]"),
+							sep = "|"),
 						verbose = 0
 		)
-		fhir_collapse(d, columns = "name.given", sep = sep, brackets = brackets)
+		fhir_collapse(d, columns = "name.given", sep = "|", brackets = c("[", "]"))
 		}, style = "json2")
 
 	})