+++ NEWS.md
@@ -3,6 +3,8 @@
 ## Bug fixes
 - Errors when removing tags in `fhir_search()` (e.g. with `rm_tag = "div"`) now get caught and converted to a warning.
 
+- When the server is down, `fhir_search()` will now actually try again as specified in the argument `delay_between_attempts` instead of throwing an error at the first attempt.
+
 ## New functions
 - `fhir_is_empty()` checks if a bundle or bundlelist is empty
 - `fhir_collapse()` collapses multiple entries that belong to the same higher level FHIR element and should be kept together, e.g. `name.given` or `address.line`