+++ NEWS.md
@@ -1,9 +1,21 @@
-# fhircrackr 2.2.1
+# fhircrackr 2.3.0
 
 ## Bug fixes
 - Errors when removing tags in `fhir_search()` (e.g. with `rm_tag = "div"`) now get caught and converted to a warning.
+- `fhir_crack() doesn't crash anymore when some of the bundles don't contain the searched resources.
 
+## New functionalities
+
+- New Argument `stop_on_error` in `fhir_search()` explicitly controls error handling for server requests: 
+	- stop_on_error = 0: Issues warning and retries requests as defined by delay_between_attempts. Returns empty
+	bundle_list if all attempts fail
+
+	- stop_on_error = 1: Stops immediately on any error
+
+	- ⁠stop_on_error = <vector of HTTP codes>⁠: Stops on specified codes (e.g. c(401, 404)), warns for others
+	
 ## New functions
+- `fhir_is_empty()` checks if a bundle or bundlelist is empty
 - `fhir_collapse()` collapses multiple entries that belong to the same higher level FHIR element and should be kept together, e.g. `name.given` or `address.line`
 
 - `fhir_melt_all()` divides all multiple entries into separate rows by repeatedly calling `fhir_melt()`.